pyproject.toml
[project]
name = "panacloud-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
